+++ src/main/java/firok/mtim/util/Keys.java
@@ -23,20 +23,48 @@ public interface Keys
 	String nameGlass= "glass";
 	String nameConsolidatedGlass= "consolidated_glass";
 	String nameBrokenIce="broken_ice";
+	String nameImmersedSilver="immersed_silver";
+	String nameMithril="mithril";
+	String nameAdamantine="adamantine";
+	String nameBlackrock="blackrock";
+	String nameInertWitherium="inert_witherium";
+	String nameWitherium="witherium";
 
 	String moltenRoyalAlloy=prefMolten+nameRoyalAlloy;
 	String moltenStellar=prefMolten+nameStellar;
 	String moltenCinnabar=prefMolten+nameCinnabar;
+	String moltenImmersedSilver=prefMolten+nameImmersedSilver;
+	String moltenMithril=prefMolten+nameMithril;
+	String moltenAdamantine=prefMolten+nameAdamantine;
+	String moltenInertWitherium=prefMolten+nameInertWitherium;
+	String moltenWitherium=prefMolten+nameWitherium;
 
 	String blockStellar=prefBlock+nameStellar;
 	String blockInkPowder=prefBlock+nameInkPowder;
 	String blockCinnabar=prefBlock+nameCinnabar;
+	String blockImmersedSilver=prefBlock+nameImmersedSilver;
+	String blockMithril=prefBlock+nameMithril;
+	String blockAdamantine=prefBlock+nameAdamantine;
+	String blockBlackrock=prefBlock+nameBlackrock;
+	String blockInertWitherium=prefBlock+nameInertWitherium;
+	String blockWitherium=prefBlock+nameWitherium;
 
 	String oreCinnabar=prefOre+nameCinnabar;
 	String oreInkPowder=prefOre+nameInkPowder;
+	String oreImmersedSilver=prefOre+nameImmersedSilver;
+	String oreMithril=prefOre+nameMithril;
+	String oreAdamantine=prefOre+nameAdamantine;
+	String oreBlackrock=prefOre+nameBlackrock;
+	String oreInertWitherium=prefOre+nameInertWitherium;
+	String oreWitherium=prefOre+nameWitherium;
 
 	String ingotStellar=prefIngot+nameStellar;
 	String ingotRoyalAlloy=prefIngot+nameRoyalAlloy;
+	String ingotImmersedSilver=prefIngot+nameImmersedSilver;
+	String ingotMithril=prefIngot+nameMithril;
+	String ingotAdamantine=prefIngot+nameAdamantine;
+	String ingotInertWitherium=prefIngot+nameInertWitherium;
+	String ingotWitherium=prefIngot+nameWitherium;
 
 	int colorStellar=Colors.Tomato;
 	int colorRoyalAlloy=Colors.Yellow;
@@ -46,6 +74,12 @@ public interface Keys
 	int colorGlass=Colors.WhiteSmoke;
 	int colorConsolidatedGlass=Colors.WhiteSmoke;
 	int colorBrokenIce=Colors.Snow;
+	int colorImmersedSilver=Colors.Silver;
+	int colorMithril=Colors.Silver;
+	int colorAdamantine=Colors.BlanchedAlmond;
+	int colorBlackrock=Colors.DarkRed;
+	int colorInertWitherium=Colors.DarkGray;
+	int colorWitherium=Colors.DimGray;
 
 	String nameTraitIcy="icy";
 	String nameTraitLuxurious="luxurious";